requirements.txt
diff-match-patch==20121119
Django==1.9.5
django-analytical==2.0.0
django-auth-ldap==1.2.7
django-countries==3.3
django-custom-user==0.5
django-filebrowser==3.6.4
django-grappelli==2.8.1
django-import-export==0.4.5
django-ipware==1.1.3
django-ratelimit==0.6.0
django-registration==2.0.4
django-shortcuts==1.5
django-simple-email-confirmation==0.12
funcsigs==0.4
#functools==0.5
mock==1.3.0
pbr==1.6.0
Pillow==3.1.1
python-ldap==2.4.25
six==1.9.0
tablib==0.11.2
win-inet-pton==1.0.1
